User token's Verification

+++ src/Controller/AuthorController.php
@@ -9,22 +9,38 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/author')]
 class AuthorController extends AbstractController
 {
     #[Route('/', name: 'app_author_index', methods: ['GET'])]
-    public function index(AuthorRepository $authorRepository): Response
+    public function index(AuthorRepository $authorRepository, TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         return $this->render('author/index.html.twig', [
             'authors' => $authorRepository->findAll(),
         ]);
     }
+    
 
     #[Route('/new', name: 'app_author_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $author = new Author();
         $form = $this->createForm(AuthorType::class, $author);
         $form->handleRequest($request);
@@ -43,16 +59,30 @@ class AuthorController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_author_show', methods: ['GET'])]
-    public function show(Author $author): Response
+    public function show(Author $author, TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         return $this->render('author/show.html.twig', [
             'author' => $author,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_author_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Author $author, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Author $author, EntityManagerInterface $entityManager, TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $form = $this->createForm(AuthorType::class, $author);
         $form->handleRequest($request);
 
@@ -69,8 +99,15 @@ class AuthorController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_author_delete', methods: ['POST'])]
-    public function delete(Request $request, Author $author, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Author $author, EntityManagerInterface $entityManager, TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         if ($this->isCsrfTokenValid('delete'.$author->getId(), $request->request->get('_token'))) {
             $entityManager->remove($author);
             $entityManager->flush();
