+++ src/Controller/CategorieController.php
@@ -9,14 +9,23 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 use Symfony\Component\Routing\Annotation\Route;
 
 #[Route('/categorie')]
 class CategorieController extends AbstractController
 {
     #[Route('/', name: 'app_categorie_index', methods: ['GET'])]
-    public function index(CategorieRepository $categorieRepository): Response
+    public function index(CategorieRepository $categorieRepository,TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
+        
         flash()->addInfo("Retriving all the category");
         return $this->render('categorie/index.html.twig', [
             'categories' => $categorieRepository->findAll(),
@@ -24,8 +33,15 @@ class CategorieController extends AbstractController
     }
 
     #[Route('/new', name: 'app_categorie_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $categorie = new Categorie();
         $form = $this->createForm(CategorieType::class, $categorie);
         $form->handleRequest($request);
@@ -44,16 +60,30 @@ class CategorieController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_categorie_show', methods: ['GET'])]
-    public function show(Categorie $categorie): Response
+    public function show(Categorie $categorie,TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         return $this->render('categorie/show.html.twig', [
             'categorie' => $categorie,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_categorie_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Categorie $categorie, EntityManagerInterface $entityManager): Response
+    public function edit(Request $request, Categorie $categorie, EntityManagerInterface $entityManager,TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         $form = $this->createForm(CategorieType::class, $categorie);
         $form->handleRequest($request);
 
@@ -71,8 +101,15 @@ class CategorieController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_categorie_delete', methods: ['POST'])]
-    public function delete(Request $request, Categorie $categorie, EntityManagerInterface $entityManager): Response
+    public function delete(Request $request, Categorie $categorie, EntityManagerInterface $entityManager,TokenStorageInterface $tokenStorage): Response
     {
+        $token = $tokenStorage->getToken();
+        if (!$token || !$token->getUser()) {
+            $this->addFlash('error', 'You need to be logged in to access this resource');
+            return $this->redirectToRoute('app_login');
+        }
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         if ($this->isCsrfTokenValid('delete'.$categorie->getId(), $request->request->get('_token'))) {
             $entityManager->remove($categorie);
             $entityManager->flush();
